requirements.txt
certifi==2022.12.7 ; python_version >= "3.11" and python_version < "4"
charset-normalizer==3.0.1 ; python_version >= "3.11" and python_version < "4"
idna==3.4 ; python_version >= "3.11" and python_version < "4"
markdown==3.4.1 ; python_version >= "3.11" and python_version < "4.0"
python-frontmatter==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.11" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.11" and python_version < "4"
urllib3==1.26.14 ; python_version >= "3.11" and python_version < "4"
